Split std into submodules
