+++ sjsonnet/test/src-jvm-native/sjsonnet/BaseFileTests.scala
@@ -8,7 +8,7 @@ abstract class BaseFileTests extends TestSuite {
   val workspaceRoot = sys.env.get("MILL_WORKSPACE_ROOT").map(os.Path(_)).getOrElse(os.pwd)
   val testSuiteRoot: os.Path = workspaceRoot / "sjsonnet" / "test" / "resources"
   private val stderr = new StringBuffer()
-  private val std = new Std(
+  private val std = new sjsonnet.stdlib.StdLibModule(
     nativeFunctions = Map(
       "jsonToString" -> new Val.Builtin1("jsonToString", "x") {
         override def evalRhs(arg1: Lazy, ev: EvalScope, pos: Position): Val = {
@@ -59,7 +59,7 @@ abstract class BaseFileTests extends TestSuite {
         }
       },
       settings = new Settings(),
-      std = std.Std
+      std = std.module
     )
     interp.interpret(os.read(p), OsPath(p))
   }
