+++ sjsonnet/test/src-jvm-native/sjsonnet/stdlib/NativeGzipTests.scala
@@ -1,6 +1,5 @@
 package sjsonnet.stdlib
 
-import sjsonnet.Std
 import sjsonnet.TestUtils.eval
 import utest._
 
@@ -8,7 +7,8 @@ object NativeGzipTests extends TestSuite {
   private val isScalaNative: Boolean = System.getenv("SCALANATIVE_THREAD_STACK_SIZE") != null
   private val javaVersion: Double =
     java.lang.Double.parseDouble(System.getProperty("java.specification.version"))
-  private val std = new Std(nativeFunctions = Map() ++ new NativeGzip().functions)
+  private val std =
+    new sjsonnet.stdlib.StdLibModule(nativeFunctions = Map() ++ new NativeGzip().functions)
 
   def isJava16OrLater: Boolean = {
     if (isScalaNative)
