+++ sjsonnet/test/src-jvm/sjsonnet/Example.java
@@ -13,7 +13,7 @@ public class Example {
             os.package$.MODULE$.pwd(),
             scala.None$.empty(),
             scala.None$.empty(),
-            new sjsonnet.Std(Map$.MODULE$.empty()).Std()
+            new sjsonnet.stdlib.StdLibModule(Map$.MODULE$.empty(), Map$.MODULE$.empty()).module()
         );
     }
 }
